new refactor that works
